register page and access to the specific user

+++ my_proj/base/views.py
@@ -4,7 +4,7 @@ from django.http import HttpResponse
 from django.db.models import Q
 from django.contrib.auth import login, authenticate, logout
 from django.contrib.auth.models import User
-
+from django.contrib.auth.decorators import login_required
 from .forms import RoomForm
 # Create your views here.
 from .models import Room, Topic
@@ -15,7 +15,11 @@ from .models import Room, Topic
 #     {'id': 4, 'name': '4 data'}
 # ]
 
+
 def loginPage(request):
+    page = 'login'
+    if request.user.is_authenticated:
+        return redirect('home')
     if request.method == 'POST':
         username = request.POST.get('username')
         password = request.POST.get('password')
@@ -27,9 +31,18 @@ def loginPage(request):
         if user is not None:
             login(request, user)
             return redirect('home')
-    context = {}
+        else:
+            messages.error(request, 'username or password does not exist')
+    context = {'page': page}
     return render(request,"base/login_page.html",context)
 
+def  logoutUser(request):
+    logout(request)
+    return redirect('home')
+
+def registerPage(request):
+    page = 'register'
+    return render(request,"base/login_page.html",)
 def home(request):
     q = request.GET.get('q') if request.GET.get('q') else ''
     rooms = Room.objects.filter(
@@ -50,7 +63,7 @@ def room(request,pk):
     #         room =i
     context = {'room':room}
     return render(request, 'base/room.html',context)
-
+# create a new room 
 def createRoom(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -62,10 +75,13 @@ def createRoom(request):
     context = {'form': form}
     return render(request,"base/room_form.html",context)
 
+# update the authorized room
+@login_required(login_url='login')
 def updateRoom(request,pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
-
+    if request.user !=room.host:
+        return HttpResponse('You are not allowed here!!!')
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
         if form.is_valid():
@@ -74,8 +90,13 @@ def updateRoom(request,pk):
     context = {'form': form}
     return render(request,"base/room_form.html",context)
 
+
+# delete an authorized room
+@login_required(login_url='login')
 def delete_room(request,pk):
     room = Room.objects.get(id=pk)
+    if request.user !=room.host:
+        return HttpResponse('You are not allowed here!!!')
     if request.method =='POST':
         room.delete()
         return redirect('home')
